Fix BufferedOpusEncoderPlayable cutting off playback too soon
Previously the signal for when BufferedOpusEncoderPlayable was
completed, was when the encoding process finished. However considering
we buffer the encoding, this would cause the playable to complete 4096
frames before it was actually done. We now properly just enqueue a None
(aka complete signal) when encoding is done, so that playback will
complete fully.