[bugfix] roles in GuildRoleCreate have guild_id
Previously guild_id sat outside of the object which created a
semi-invalid role object without guild information. This would be saved
in the guild state.

One option for fixing this would just to be fixing the guild_id of the
role within the state, however I think it's responible to expect roles
passed from the event to have valid context/guild attributes.