[typing] allow Field to have a default of None
Previously we used None as the sentinel for determining that a user had
not specified a default value for a field. This meant that a user could
never select `None` as the default value. We now properly use UNSET to
determine if a default was passed, this aligns better with the code as
it stands given that if no default is passed we'll use UNSET anyway.