Clear voice states when a new session connects (fixes #100, closes #109)
Discord is inconsistent here and doesn't emit a VoiceStateUpdate for the
old session_id when a user connects to voice from a different device
(thus disconnecting the first device). To avoid corrupting our voice
state cache, we need to remove any of the users old sessions (they
should have at most one).

I included tests on this one because its esoteric and undocumented, so
there is a high chance I'll break it in the future.