Remove references to deprecated `trace` field
This will not undergo a regualar deprecation cycle since the schedule
provided by discord (see discordapp/discord-api-docs#967) is less than
the amount of time it would take to fully deprecate this field. Instead
we will just consider this a v1 breaking change.

Clients on previous versions should continue to work regardless of
Discord's planned changes for this field, although the data provided
from the field may be invalid or empty.